fix the issue in update content component

+++ screens/Content/SubScreens/ViewContentSubPage.js
@@ -7,6 +7,7 @@ import {
   ScrollView,
   Dimensions,
   ActivityIndicator,
+  Alert,
 } from "react-native";
 import React, { useEffect, useState } from "react";
 import { useNavigation } from "@react-navigation/native";
@@ -17,8 +18,25 @@ const ViewContentSubPage = ({ content, isAuthor }) => {
   const navigation = useNavigation();
 
   const deleteContent = () => {
-    deleteDoc(doc(db, "Content", content.id));
-    navigation.navigate("MyContentList");
+    Alert.alert(
+      "Delete Content",
+      "Are you sure you want to delete this content?",
+      [
+        {
+          text: "Cancel",
+          onPress: () => console.log("Cancel Pressed"),
+          style: "cancel",
+        },
+        {
+          text: "OK",
+          onPress: () => {
+            deleteDoc(doc(db, "Content", content.id));
+            navigation.navigate("MyContentList");
+          },
+        },
+      ],
+      { cancelable: false }
+    );
   };
 
   return (
